\name{BinBam.gcCorrect}
\alias{BinBam.gcCorrect}
\title{BinBam.gcCorrect}
\usage{
BinBam.gcCorrect(bam.file, index = bam.file, gcContent = NULL)
}
\arguments{
  \item{bam.file}{name of bamfile}

  \item{index}{index of bam file}

  \item{gcContent}{name of the gccontent data}
}
\description{
Given a list of bam files, this function writes the output
to a text file after binning and doing gc Correction NOTE:
Not for production use
}

