% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BinBam}
\alias{BinBam}
\title{BinBam}
\usage{
BinBam(bam.file, index = bam.file, mask = NULL, k = 20000)
}
\arguments{
\item{bam.file}{name of bamfile}

\item{index}{index of bam file}

\item{mask}{mask file in the bed format}

\item{k}{bin size in kilobases}
}
\description{
Given a list of bam files, this function writes the output
to a text file after binning and doing gc Correction
}

