% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{rapid_date}
\alias{rapid_date}
\title{Get the acquisition date}
\usage{
rapid_date(xml)
}
\arguments{
\item{xml}{An XML document}
}
\value{
A class Date
}
\description{
Get the acquisition date
}
\examples{
library(xml2)

xml.file <- list.files(
  system.file("extdata", package="rapidr"),
  pattern = ".xml", full.names = TRUE
)

x <- read_xml(xml.file)

rapid_date(x)

}

