% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{rapid_epsg}
\alias{rapid_epsg}
\title{Get the EPSG reference coordinate}
\usage{
rapid_epsg(xml)
}
\arguments{
\item{xml}{An XML document}
}
\value{
A SpatialPoints of central coordinates from Rapideye tile
}
\description{
Get the EPSG reference coordinate
}
\examples{
library(xml2)

xml.file <- list.files(
  system.file("extdata", package="rapidr"),
  pattern = ".xml", full.names = TRUE
)

x <- read_xml(xml.file)

rapid_epsg(x)

}

