% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{rapid_hour}
\alias{rapid_hour}
\title{Get the acquisition hour}
\usage{
rapid_hour(xml)
}
\arguments{
\item{xml}{An XML document}
}
\value{
A decimal hour
}
\description{
Get the acquisition hour
}
\examples{
library(xml2)

xml.file <- list.files(
  system.file("extdata", package="rapidr"),
  pattern = ".xml", full.names = TRUE
)

x <- read_xml(xml.file)

rapid_hour(x)

}

