% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflectance.R
\name{rapid_reflec}
\alias{rapid_reflec}
\title{Convert the Digital Number from raw image to Reflectance}
\usage{
rapid_reflec(raster, xml)
}
\arguments{
\item{raster}{A RasterBrick object}

\item{xml}{An XML document}
}
\value{
A RasterBrick object
}
\description{
Convert the Digital Number from raw image to Reflectance
}
\examples{
library(raster)
library(xml2)

xml.file <- list.files(
  system.file("extdata", package="rapidr"),
  pattern = ".xml", full.names = TRUE
)
raster.file <- list.files(
  system.file("extdata", package="rapidr"),
  pattern = ".tif", full.names = TRUE
)
x <- read_xml(xml.file)
r <- raster::brick(raster.file)

rapid_reflec(r, x)

}

