% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{rapid_sf}
\alias{rapid_sf}
\title{Get the Scale Factor for all bands}
\usage{
rapid_sf(xml)
}
\arguments{
\item{xml}{An XML document}
}
\value{
A vector contain the scale factor of each 5 bands
}
\description{
Get the Scale Factor for all bands
}
\examples{
library(xml2)

xml.file <- list.files(
  system.file("extdata", package="rapidr"),
  pattern = ".xml", full.names = TRUE
)

x <- read_xml(xml.file)

rapid_sf(x)

}

